pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt-magic"
version = "0.2.0"
description = "A Jupyter/IPython magic command for generating code using GPT."
authors = [{name = "BIFF", email = "info@biff.ai"}]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
keywords = ["jupyter", "ipython", "gpt", "ai"]
dependencies = [
    "openai",
    "requests"
]
requires-python = ">=3.6"

[options]
packages = "find:"

[project.optional-dependencies]
test = ["pytest", "pytest-mock", "requests-mock", "notebook", "python-dotenv"]
dev = ["pytest", "pytest-mock", "requests-mock", "notebook", "python-dotenv"]

[project.urls]
Homepage = "https://biff.ai/bif-0003"
